initial commitment, version 0.2.0:
This version comes from an earlier project, and contains basically the
same code as the other project. The main difference is that earlier,
task_error, task and task_manager have been separate modules.

For this project, they are bundled together in their own 'tasks' crate,
as they all handle the same domain.